% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_flows.R
\name{create_dist_df}
\alias{create_dist_df}
\title{Create a data frame of distances between tribes}
\usage{
create_dist_df(model, threshold)
}
\arguments{
\item{model}{Object of the class \code{prehistorik_model}.}

\item{threshold}{Distance in meters after which distances are not calculated.}
}
\value{
A data frame containing all the pairwise distances between tribes.
}
\description{
Given a \code{prehisotrik_model}, this function calculate the pairwise distances in meters between all the exisitnf tribes.
}
