% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_flows.R
\name{create_gf_df}
\alias{create_gf_df}
\title{Create pairwise gene flow rates}
\usage{
create_gf_df(dist_df, t_cutoff, gf_fun, ...)
}
\arguments{
\item{dist_df}{Object of the type \code{prehistorik_dist_df} as returned by the function \code{create_dist_df}.}

\item{t_cutoff}{Time of coexistence of tribes below which gene flow cannot happen.}

\item{gf_fun}{Function that takes as input the distance between tribes to calculate the proportion of gene flow.}

\item{...}{}
}
\value{
A data frame of containing the gene flow rates between each pair of tribes.
}
\description{
Create pairwise gene flow rates
}
