% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_generation.R
\name{create_prehistorik_model}
\alias{create_prehistorik_model}
\title{Create a prehistorir metapopulation model}
\usage{
create_prehistorik_model(
  start_pop_mat,
  map_mat,
  b = 1,
  d = 0,
  t_start = 1e+06,
  t_stop = 0,
  live_map_3d,
  generation_time = 30,
  jolly = NULL,
  N = NULL,
  radius = NULL,
  kernel = NULL,
  n_max = NULL,
  ...
)
}
\arguments{
\item{start_pop_mat}{Matrix corresponding to the starting populations.}

\item{map_mat}{Object of the \code{prehistorik_map} class generated by the function \code{get_map_as_matrix}.}

\item{b}{Birth rate. Probability that a tribe give rise to a population.}

\item{d}{Death rate. Probability that a tribe dies.}

\item{t_start}{Time at which the metapopulation model starts.}

\item{t_stop}{Time at which the metapopulation model stops.}

\item{live_map_3d}{Liveability map in 3D, where each layer correspond to one generation of the model.}

\item{generation_time}{Generation time in years.}

\item{jolly}{Event ...}

\item{N}{Either a function or a constant, corresponding to the number of individuals for each new tribe.}

\item{radius}{Radius within which a new tribe can be spawned.}

\item{kernel}{(optional) Kernel used by the convolution to establish the number of tribes in a neighbourhood. If not provided a standard kernel will be generated.}

\item{n_max}{Maximum number of tribes that can simultaneously live in the model.}

\item{...}{}
}
\value{
An object of the class \code{prehistorik_model}, which corresponds to a data frame containing the history of the metapopulation.
}
\description{
Function to create a spatial metapopulation model that undergoes extinction and recolonization dynamics according to a
given birth and death rate.
}
