% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_matrix.R
\name{generate_kernel}
\alias{generate_kernel}
\title{Generate a 1s kernel}
\usage{
generate_kernel(map_mat, radius)
}
\arguments{
\item{map_mat}{An object of the class \code{prehistorik_map}.}

\item{radius}{Radius in meters for the kernel.}
}
\value{
A matrix
}
\description{
Generate a 1s kernel
}
