% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_matrix.R
\name{get_xy_from_latlon}
\alias{get_xy_from_latlon}
\title{Title}
\usage{
get_xy_from_latlon(lat, lon, map_mat)
}
\arguments{
\item{lat}{aa}

\item{lon}{aa}

\item{map_mat}{aa}
}
\value{
aa
}
\description{
Title
}
