% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_flows.R
\name{plot_gf_as_graph}
\alias{plot_gf_as_graph}
\title{Plot gene flows as graph}
\usage{
plot_gf_as_graph(model, time, threshold, lat_range = NULL, lon_range = NULL)
}
\arguments{
\item{model}{Object of the class \code{prehistorik_model}.}

\item{time}{Time at which to extract the gene flows}

\item{threshold}{Distance in meters at which to not show gene flows.}

\item{lat_range}{Range of latitudes to plot.}

\item{lon_range}{Range of longitudes to plot.}
}
\value{
Plot of gene flows.
}
\description{
Plot gene flows as graph
}
