% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_matrix.R
\name{plot_map_matrix}
\alias{plot_map_matrix}
\title{Plot prehistorik map}
\usage{
plot_map_matrix(map_mat, plotly = FALSE)
}
\arguments{
\item{map_mat}{Object of the class \code{prehistorik_map} to be plotted.}

\item{plotly}{Should the plot be interactive?}
}
\value{
Plot.
}
\description{
Plot a prehistorik map
}
