% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_matrix.R
\name{set_map_mat_values}
\alias{set_map_mat_values}
\title{Set entries of a \code{prehistorik_map}}
\usage{
set_map_mat_values(map_mat, value, x, y, subset = FALSE, na = FALSE)
}
\arguments{
\item{map_mat}{Object of the class \code{prehistorik_map}.}

\item{value}{Value to insert at the given coordinates.}

\item{x}{x coordinate(s) to modify.}

\item{y}{y coordinate(s) to modify.}

\item{subset}{Should the (x,y) coordinates be treated as a list of points or as an area?}

\item{na}{If the existing value is NA, should we modify it?}
}
\value{
\code{map} given as input, but with the entries modified.
}
\description{
Function to modify entries of a \code{prehistorik_map}.
}
